requirements.txt
Flask==1.1.4
markupsafe==2.0.1
gunicorn==20.1.0
convertdate==2.1.2
lunarcalendar==0.0.9
holidays==0.10.3
cython==0.29.21
Cython>=0.22
pystan==2.18
numpy==1.19.1
pandas==1.1.1
cmdstanpy==0.4
setuptools-git==1.2
pillow==6.2.0
kiwisolver==1.0.1
cycler==0.10
google-cloud-bigquery==2.13.1
google-cloud-bigquery-storage==2.3.0
pyarrow==3.0.0
plotly==4.14.3
